problem4.1 half code
1) Write a Menu driven C program to accomplish the following functionalities in doubly
linked list.
a) Create a doubly linked list.
b) Display the elements of a doubly linked list.
c) Insert a node at the beginning of a doubly linked list.
d) Insert a node at the end of a doubly linked list
g) Delete a node from the beginning of a doubly linked list.
h) Delete a node from the end of a doubly linked list.